[Gold IV] Title: 회사 문화 1, Time: 44 ms, Memory: 15616 KB -BaekjoonHub

+++ 백준/Gold/14267. 회사 문화 1/회사 문화 1.cc
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-long long dp[100001]={0};
+int dp[100001]={0};
 
 void DFS(int node , vector<vector<int>> &graph , vector<bool> &visit)
 {
